requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
numpy==1.24.3
pandas==2.1.3
scipy==1.11.4
shap==0.43.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
asyncpg==0.29.0

# Caching
redis==5.0.1
hiredis==2.2.3

# ML Tracking
mlflow==2.9.2

# API & Async
httpx==0.25.2
python-multipart==0.0.6

# Slack SDK
slack-sdk==3.33.4

# Environment & Config
python-dotenv==1.0.0
pyyaml==6.0.1

# OpenAI & LangChain
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.2
langchain-core==0.1.10

# Utilities
joblib==1.3.2
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1

seaborn
python-dotenv